+++ packages/client/src/Components/Mapping/helpers/colorFor.js
